fix(web): rename component and improve documentation
Renamed the component to make it more agnostic. While it defaults to
"skip to content," it can be configured to skip to navigation or
other sections as needed.
