+++ web/src/components/layout/Layout.tsx
@@ -26,7 +26,7 @@ import { Masthead, Page, PageProps } from "@patternfly/react-core";
 import { Questions } from "~/components/questions";
 import Header, { HeaderProps } from "~/components/layout/Header";
 import { Loading, Sidebar } from "~/components/layout";
-import { IssuesDrawer, SkipToContentLink } from "~/components/core";
+import { IssuesDrawer, SkipTo } from "~/components/core";
 import { ROOT } from "~/routes/paths";
 import { agamaWidthBreakpoints, getBreakpoint } from "~/utils";
 
@@ -100,7 +100,7 @@ const Layout = ({
     // FIXME: render an empty Masthead instead of nothing, in order to have
     // everything working as designed by PatternFly (there are some CSS rules
     // that expect the masthead to be there :shrug:)
-    pageProps.masthead = <Masthead>{mountSkipToContent && <SkipToContentLink />}</Masthead>;
+    pageProps.masthead = <Masthead>{mountSkipToContent && <SkipTo />}</Masthead>;
   }
 
   return (
